Decouple `BloggingPromptSettings` file from personalization service
This will allow us to move `BloggingPromptSettings` to WordPressData as
part of #24165 without dragging over the
`BlogDashboardPersonalizationService` dependency.
