+++ WordPress/Classes/Models/BloggingPromptSettings+CoreDataClass.swift
@@ -27,17 +27,6 @@ public class BloggingPromptSettings: NSManagedObject {
         return try context.fetch(fetchRequest).first
     }
 
-    func configure(with remoteSettings: RemoteBloggingPromptsSettings, siteID: Int32, context: NSManagedObjectContext) {
-        self.siteID = siteID
-        self.promptCardEnabled = remoteSettings.promptCardEnabled
-        self.reminderTime = remoteSettings.reminderTime
-        self.promptRemindersEnabled = remoteSettings.promptRemindersEnabled
-        self.isPotentialBloggingSite = remoteSettings.isPotentialBloggingSite
-        updatePromptSettingsIfNecessary(siteID: Int(siteID), enabled: isPotentialBloggingSite)
-        self.reminderDays = reminderDays ?? BloggingPromptSettingsReminderDays(context: context)
-        reminderDays?.configure(with: remoteSettings.reminderDays)
-    }
-
     func reminderTimeDate() -> Date? {
         guard let reminderTime else {
             return nil
@@ -45,15 +34,6 @@ public class BloggingPromptSettings: NSManagedObject {
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "HH.mm"
         return dateFormatter.date(from: reminderTime)
-    }
-
-    private func updatePromptSettingsIfNecessary(siteID: Int, enabled: Bool) {
-        DispatchQueue.main.async {
-            let service = BlogDashboardPersonalizationService(siteID: siteID)
-            if !service.hasPreference(for: .prompts) {
-                service.setEnabled(enabled, for: .prompts)
-            }
-        }
     }
 }
